PCHR-3084: Move test object in separate module
